Added first implementation of the config module
The config module defines the PySIMS configuration.

The module provides the interface for handling the program's configuration.
The configuration constants of PySIMS are also stored here.

The `ConfigManager` class based on `pydantic.BaseSettings` serves
as the main interface to process the PySIMS configuration.

The unit tests for the config module now pass.